Fix: change doubles to decimals in test projects

+++ Tests/AlgorithmsTest/EloRatingCalculatorTest.cs
@@ -12,13 +12,13 @@ namespace Catmash.Tests.Algorithms
         public void ComputeExpectation_ratingsGiven_ShouldReturnExpectation()
         {
             // Arrange
-            double ratingA = 123.4;
-            double ratingB = 865.23;
-            double expectation = 1 / (1 + Math.Pow(10, (ratingB - ratingA) / 400));
+            decimal ratingA = 123.4M;
+            decimal ratingB = 865.23M;
+            decimal expectation = (decimal)(1 / (1 + Math.Pow(10, (double)((ratingB - ratingA) / 400))));
             EloRatingCalculator ratingCalculator = new EloRatingCalculator();
 
             // Act
-            double computedExpectation = ratingCalculator.ComputeExpectation(ratingA, ratingB);
+            decimal computedExpectation = ratingCalculator.ComputeExpectation(ratingA, ratingB);
 
             Assert.Equal(expectation, computedExpectation);
         }
@@ -27,16 +27,16 @@ namespace Catmash.Tests.Algorithms
         public void ComputeRating_CurrentRatingAndExpectedAndActualScoresAndKFactorGiven_ShouldReturnNewRating()
         {
             // Arrange
-            double currentRating = 757.34;
-            double expectedScore = 0.74;
-            double actualScore = 1;
-            double kFactor = 24;
+            decimal currentRating = 757.34M;
+            decimal expectedScore = 0.74M;
+            decimal actualScore = 1;
+            decimal kFactor = 24;
 
-            double newRating = currentRating + kFactor * (actualScore - expectedScore);
+            decimal newRating = currentRating + kFactor * (actualScore - expectedScore);
             EloRatingCalculator ratingCalculator = new EloRatingCalculator();
 
             // Act
-            double computedRating = ratingCalculator.ComputeRating(
+            decimal computedRating = ratingCalculator.ComputeRating(
                 currentRating,
                 expectedScore,
                 actualScore,
